requirements.txt
datasets
matplotlib
numpy
openai
pandas
python-dotenv
seaborn
tqdm
